+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Will now log the extension version to the console on startup
 - The get help button now shows the top committer name instead of "on Slack"
 - Fixed an error where the command would break even when results were fetched
+- Fixed some instances of the get help on Slack button not working
 - Will select whole file if no selection is made
 
 ## [1.1.3]